Discovered the reason why progress was not being reported.
"The WriteObject and WriteError methods cannot be called from outside the overrides of the BeginProcessing, ProcessRecord, and EndProcessing methods, and they can only be called from within the same thread".

Solution is to use an AsyncOperation and Windows Message Pump (borrowed from Windows.Forms) and invoke the Write* methods on the Pipeline Thread.